requirements.txt
streamlit==1.39.0
yfinance==0.2.40
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.1
tensorflow==2.16.2  # Replaced tensorflow-macos with tensorflow
joblib==1.4.2
requests==2.32.3
websocket-client==1.8.0
ib-insync==0.9.70
jinja2==3.1.6
prometheus-client==0.20.0
pytest==8.3.5
pytest-asyncio==0.25.3